refactor pecPorts in pipes-shell-2
